refactor: rename customer update API file
